Update with colormap
