+++ platformio.ini
@@ -10,7 +10,6 @@
 
 [platformio]
 default_envs = xiaoblesense
-monitor_speed = 115200
 [env]
 lib_deps_external = 
 	arduino-libraries/ArduinoBLE@^1.3.1
@@ -33,7 +32,7 @@ board = seeed_xiao_esp32c3
 framework = arduino
 lib_deps = 
 	${env.lib_deps_external}
-
+build_flags = -std=c++17
 
 [env:m5stack-atom]
 platform = espressif32
